requirements.txt
Django==2.1
djangorestframework==3.7.7
google-cloud-logging==1.7.0